Remove FLAccordion
