--- packages/components/src/lib/primitives/FLAccordion/index.module.css
@@ -1,7 +1,0 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
