--- packages/components/src/lib/primitives/FLAccordion/index.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FLAccordion from './index';
-
-describe('FLAccordion', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FLAccordion />);
-    expect(baseElement).toBeTruthy();
-  });
-});
