--- packages/components/src/lib/primitives/FLAccordion/index.stories.tsx
@@ -1,12 +1,0 @@
-import type { Meta } from '@storybook/react';
-import { FLAccordion } from './index';
-
-const Story: Meta<typeof FLAccordion> = {
-  component: FLAccordion,
-  title: 'FLAccordion',
-};
-export default Story;
-
-export const Primary = {
-  args: {},
-};
