+++ packages/components/src/lib/primitives/index.ts
@@ -1,4 +1,3 @@
 export * from './FLRadioGroup'
 export * from './FLTooltip'
-export * from './FLAccordion'
 export * from './FLMultiAccordion'
